#251 moved custom query to separate class - SRP principle
